refactor(models): split diplomacy.dart into concern libraries
Keep public barrel exports stable while shrinking the diplomacy host under
the models 500 NCL cap (Refs #4068 Slice B).
